fix TeamBoard Latitude Longitude

+++ src/main/java/com/springboot/publicplace/dto/request/TeamBoardRequestDto.java
@@ -11,6 +11,6 @@ public class TeamBoardRequestDto {
     private String content;
     private String Image;
     private String matchLocation;
-    private Long longitude;
-    private Long latitude;
+    private Double longitude;
+    private Double latitude;
 }
